Consensus term stores extend memory variable termstores.
